Rewrite start.
