--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ['react', 'es2015', 'stage-1'],
-  "plugins": ['add-module-exports']
-}
