+++ package.json
@@ -5,38 +5,42 @@
   "version": "0.0.3",
   "author": "Jason Stallings <jason.stallin.gs>",
   "dependencies": {
+    "babel-plugin-styled-components": "^1.5.1",
     "bluebird": "^3.5.1",
     "fetch-everywhere": "^1.0.5",
     "firebase-admin": "^5.4.2",
     "firebase-functions": "^0.7.1",
-    "gatsby": "^1.9.68",
-    "gatsby-image": "^1.0.4",
-    "gatsby-link": "latest",
-    "gatsby-plugin-google-analytics": "^1.0.8",
-    "gatsby-plugin-offline": "^1.0.10",
-    "gatsby-plugin-react-helmet": "^1.0.7",
+    "gatsby": "next",
+    "gatsby-image": "next",
+    "gatsby-plugin-google-analytics": "next",
+    "gatsby-plugin-react-helmet": "next",
     "gatsby-plugin-sentry": "^0.1.0",
-    "gatsby-plugin-sharp": "^1.6.7",
-    "gatsby-plugin-styled-components": "latest",
-    "gatsby-plugin-typography": "^1.7.9",
-    "gatsby-source-filesystem": "^1.5.39",
-    "gatsby-source-wordpress": "^2.0.93",
-    "gatsby-transformer-sharp": "^1.6.5",
+    "gatsby-plugin-sharp": "next",
+    "gatsby-plugin-styled-components": "next",
+    "gatsby-plugin-typography": "next",
+    "gatsby-source-filesystem": "next",
+    "gatsby-source-wordpress": "next",
+    "gatsby-transformer-sharp": "next",
     "html-to-react": "^1.3.1",
     "lodash": "^4.16.4",
     "material-ui": "^0.19.4",
     "mkdirp": "^0.5.1",
     "nodemailer": "^4.2.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-github-widget": "0.0.7",
+    "react-helmet": "^5.2.0",
     "react-icons": "^2.2.7",
     "react-lastfm-widget": "0.0.4",
     "react-mailto": "^0.4.0",
     "react-notification-system": "^0.2.15",
     "react-social-icons": "2.7.0",
     "react-syntax-highlighter": "^5.7.0",
+    "react-typography": "^0.16.13",
     "recursive-readdir": "^2.2.1",
+    "styled-components": "^3.3.3",
     "typeface-roboto": "0.0.35",
-    "typography": "^0.16.6",
+    "typography": "^0.16.17",
     "typography-theme-github": "^0.15.10",
     "whatwg-fetch": "^2.0.3",
     "window-or-global": "^1.0.1"
@@ -60,6 +64,7 @@
     "unit": "jest"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.0.0-beta.53",
     "babel-jest": "^21.2.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
