+++ src/pages/index.js
@@ -2,7 +2,9 @@ import React from 'react';
 import Helmet from 'react-helmet';
 import Img from 'gatsby-image';
 import styled from 'styled-components';
+import { graphql } from 'gatsby';
 import LastGitHubProject from '../components/LastGitHubProject';
+import Layout from '../components/Layout';
 
 const HeaderImage = styled(Img)`
   margin-bottom: 1rem;
@@ -10,7 +12,7 @@ const HeaderImage = styled(Img)`
 
 const Home = ({ data }) =>
   (
-    <div>
+    <Layout>
       <Helmet title={`${data.site.siteMetadata.title} | ${data.site.siteMetadata.subtitle}`} />
       <HeaderImage
         title="Image from https://www.pexels.com"
@@ -23,7 +25,7 @@ const Home = ({ data }) =>
         <p>Recently I spend most of my days playing with <a href="http://wordpress.org/">WordPress</a>,&nbsp;<a href="http://nodejs.org/">Node.js</a>, and <a href="https://reactjs.org">React</a>.</p>
       </div>
       <LastGitHubProject username="octalmage" />
-    </div>
+    </Layout>
   );
 
 export default Home;
@@ -36,9 +38,9 @@ export const pageQuery = graphql`
         subtitle
       }
     }
-    headerImage: imageSharp(id: { regex: "/macbook/" }) {
-      sizes(maxWidth: 1240 ) {
-        ...GatsbyImageSharpSizes
+    headerImage: imageSharp(fluid: { originalName:{ regex:"/macbook/" } }) {
+      fluid(maxWidth: 1240) {
+        tracedSVG
       }
     }
   }
