+++ src/templates/category.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-expressions */
 import React from 'react';
 import Helmet from 'react-helmet';
-import Link from 'gatsby-link';
+import { Link, graphql } from 'gatsby';
 import PostIcons from '../components/PostIcons';
 import { rhythm } from '../utils/typography';
